requirements.txt
alembic==1.0.10
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
Faker==1.0.7
Flask==1.0.3
Flask-Admin==1.5.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
mock==3.0.5
normalize==1.0.4
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
richenum==1.2.1
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.3.3
SQLAlchemy-Utils==0.33.11
stripe==2.29.4
text-unidecode==1.2
unique==0.1
urllib3==1.25.3
validators==0.13.0
Werkzeug==0.15.4
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
